+++ src/backends/rustler/trait_bridge.rs
@@ -18,7 +18,6 @@
 
 mod bridge_functions;
 mod generator;
-mod json_args;
 mod methods;
 mod native_args;
 #[cfg(test)]